[IMP] estate: completed wizard and started with OWL exercises
In the wizard task , created wizard button to add offers to multiple selected
properties simultaneously. Created a wizard model that inherits from
TransientModel and created the necessary fields and created the method to
add multiple offers by using _context.get to get the current session and the
active_ids , then created the form view and action for the wizard and added
the type="action" button to the list view of properties.

Started with OWL framework , completed few basic exercies like creating
component and passing props to it's subcomponents and also passing
callback methods to pass data from subcomponents to parent component.
Started working on the todo list.