[IMP] estate: create custom views for properties
create a custom form for properties for better visuals and to suit
business needs. create a notebook to be able to separate related fields
together in one page. add a custom search bar to better search in
relevant fields related to properties, group properties by postcode, or
filter them depending on availability or active status. set a new
suitable title for the action.

chapter-6