[IMP] estate: create estate property type logic
add a new model enabling users to specify a tag for each property.
implement menus to simplify accessing property tags represented in
settings and property tags. create access rules for the new model. add a
views file describing the related action, and list and form views.
link the model to the __init__ file for the module. add views file to
__manifest__ to be recognized and loaded. add two new fields for the
estate.property model to be able to access the buyer and seller in the
form, and search.

chapter-7