[IMP] estate: create tags to estate properties
create an estate property tag model to be added to properties specifying
more context about them. create a new menu to simplify accessing tags.
add a field to estate.property to show and add related tags to the
property. create some views and an action to suite viewing, creating and
accessing tags. add new views file to __manifest__ to be loaded when the
server starts.

chapter-7