[IMP] estate: create property offer model
create a new estate.property.offer model to manage offers for specific
properties. create list, and form views for the offer to manage how it
looks. add offer_ids field to estate.property model to be able to view
offers from inside the property form. create a new page in the notebook
form view for the estate.property for simplifying accessing the offers.
link views file to manifest to be loaded when starting the server. add
access rules for the new model. rename estate_property_tag file to
estate_property_tag_views to meet the convention.

chapter-7