[IMP] estate: add computed and onchange fields
create computed fields to estate.property to compute the total area
automatically when the user changes one of the two summed values. add a
onchange behavior for the garden toggling that changes garden_rea and
garden_orientation's values either by resetting them when toggled off
and putting a default value when toggled on. implement a computed field
and inverse method in property offer model to be able to see the
validity days when changing the deadline while updating the validity
when changing the deadline as well. change the form and list views for
offers to reflect the changes.

chapter-8