[IMP] estate: add actions to deal with offers
create two buttons for selling and canceling a property and automtaing
the process of changing the state. handle the logic of not selling a
cancelled property and not canceling a sold property. add two buttons
for offers to accept and refuse them, handle setting the buyer and
selling price for the property upon acceptance. handle the logic of not
accepting more than one offer at the same time for the same property.
implement logic for clearing buyer and selling price data upon refusing
a previously accepted offer to enable accepting another offer.

chapter-9