[IMP] estate: add constraints for data consistency
add sql constraints to estate property to ensure expected price and
selling price are positive and raise errors of that wasn't the case. add
sql constrains to estate property offer to ensure the offered price is
always positive. attach a sql constraint check on both type and tag
to ensure they are always unique and provide an error message otherwise.
implement python constrains check to estate property to check the selling price being at least 90% of the expected one.
use float utils from odoo to better deal with floats. define a couple of
constants describing the values of the threshold should be met for the
selling price and the epsilon when rounding floats in estate properties
to achieve a readable and clean code avoiding magic numbers.

chapter-10