[IMP] awesome_dashboard: completed OWL chapter 2 till 6th exercise
Learned the concept of services which U exports state or functions that are
persistent across components and provide a global state. Then used the
action service to open pre-defined actions on button click. Used a rpc call to
get data from the controller and display it. Created a new service to and used
the memoize function to  not call the rpc everytime component is mounted and
instead use the cached data. Used the loadJs function to load the external
Chart.js library and rendered a pie chart to display the order_by_size data.