[IMP] estate: add the sprinkles
add colors for tags, properties, offers to make things look better
depending on the preference and state of the record. implement
conditional appearance for buttons and fields to suit the business needs
and prevent accidental errors. add confirmation messages to cancel and
sell buttons to prevent accidental errors. create a statinfo button to
show the number of offers for that type of properties and view them
directly. add a filter domain to find properties with living area
greater than or equal to the entered number for better convenience.
enable the option of editing records inline instead of openeing the form
view when accessed from another list view.

chapter-11