[IMP] estate: apply changes using inheretance
apply ondelete method to prevent deletion of a property if its state is
not 'new' or 'cancelled'. override create method in offers to handle
preventing creating an offer with a lowr price than the maximum price
offered. extend user form view to add a new page for estate properties
related to that user using view inheritance.

chapter-12