[IMP] awesome_dashboard: completed chapter 2 of OWL
Returned reactive object from the service and then subscribed to it by using
useState to reflect the changes in data in 10 minutes and update it immediately
if the dashboard is being displayed.  Added all the dashboard files in a sub-
folder and defined it in the manifest as a bundle. Then registered awesome_
dashboard.dashboard in the lazy_components registry instead of actions
and registere another component which has lazy component component as actions
where lazy component is a intermediary component which loads the bundle when
it's requested. Made the dashboard generic and extensible by registering items
in a  category called awesome_dashboard instead of directly importing it from a
file. Created a settings dialog component by using the inbuilt dialog component
where user can select the items that they want. Created a hook to add the dialog
using the dialog service and stored the user setting preference in the local
storage so that the preference persists even after page reload.