[REF] estate: refactor methods for better process
refactor create overridden method to process handling batches of records instead of single records.
modify kanban view for the estate property to appear. use card instead of kanban-box in kanban view as it is no longer used.
move the logic of changing the state of the property to 'offer_received' to the create in offer model instead of onchange in estate_property model for better separation of concerns applying SRP.
move the constants for the offer price threshold should be met and the epsilon for floats comparison to a separate file making it central to be used all over the module instead of redefining it locally at many places.