[IMP] awesome_owl: use callback to add counters
define a new state in the playground to hold the summation of the two children counters.
bind the onChange function to increase the sum by one to define the context for the function. define the optional prop as an onChange function that is called when the child's state is changed to update the parent's sum.

chapter-1-part-6