[IMP] awesome_owl: add a todo list to playground
create a todo item component that takes a prop todo and shows its id and description in a card style.
add a todolist component that accepts an optional prop todos and keeps a state for the available todos, iterates over them and displays a list of TodoItem.
append the todolist to the playground.

chapter-1-part-7