[ADD] awesome_dashboard: making dashboard extensible and used dialogbox
learned about making our dashboard extensible. Instead of exporting a list,
we register all dashboard items in an awesome_dashboard registry. Then, we
import all the items in the Dashboard component. also learned how to make
the dashboard customizable. For that, we add a gear icon as an action button.
After clicking that button, a dialog opens. In that dialog, after clicking the
apply button, it stores the unchecked items in local storage. We then
filter out those items from our dashboard.