[IMP] awesome_owl: enable users to add tasks
add an input field that handles the event of key pressing and reacts to them by adding a new task to the todos if they're suitable.
implement addTask method in todo_list that handles the process of adding tasks and updating the state. maintain a global variable in the TodoList to handle generating new ids for tasks.

chapter-1-part-9