[ADD] product_warranty: add warranty for products task
Added a add warranty wizard which displays a list of products that are added in
the sales order and have is_warrant_available ticked. Then a warranty product
can be chose which is then added in it's own line. When the product is deleted
the added warranty for that product is deleted too.