[IMP] awesome_owl: enable removing todos
add callback prop to invoke re-renders on clicking the remove icon. change the conditional decoration to be applied only on the body of the card instead of applying it globally on the card. add the remove icon to the card header. change the header display to flex for better alignment for the elements.

chapter-1-part-12