[ADD] default_attribute: allow user to set default attribute on product category
PURPOSE

The purpose of this module is to allow users to set attributes on product
category which can be directly assigned to the product of that category as a
default attributes.

TECHNICAL DETAILS

Add new fields show_on_global_info and attributes_ids to product_category model.
And add show_on_global_info and attributes_ids fields to product_category_views
file.

Task-4606091