[IMP] awesome_dashboard: make the dashboard generic
add an items file containing the definition of each component of the dashboard items. genericly use a foreach to render eash of these items with the appropriate data in new dashboard items. create two card components to specify whether the item being renderd is a number card or a pie chart card.

chapter-2-part-9