[IMP] sale_kit_sub_products : Add  Is Kit with subproduct handling.
Added an Is Kit checkbox in the product.template model to designate kit
products.When a kit product is added to a Sales Order, a button Add Subproducts
appears in the order line. Clicking this button opens a wizard displaying the
associated subproducts, and upon confirmation, all subproducts are automatically
added to the Sales Order lines with a price of zero and set as non-editable. The
total price of the subproducts is included in the main product's price, ensuring
correct pricing and invoicing. Additionally, introduced a Print in Report
checkbox in the Sales Order, allowing subproducts to be displayed in the
Quotation, Customer Preview, and Invoice reports when selected.