[IMP] pos_salesperson: enable salesperson pos in list ad email and orderline
this commit aims to  Load hr.employee model into POS for salesperson selection.
also sales_person_id field to pos.order to store the assigned salesperson. Patch
OrderWidget to display the selected salesperson in the order
summary.Patch PosOrder to manage salesperson assignment and access.
Introduce a control button on the billing screen to select, change, or remove
the salesperson.Display salesperson information in the form and list views of
POS orders.

task-4617123