[ADD] sell_products_kit:  New Product Type in Odoo
Implemented a custom module new_product_type that introduces sub products
for a product. A boolean field is_kit is added to the product form, and
when set to True, a sub product field becomes visible, allowing users to add
other products except for the product itself. In the sale order line, a button
labeled "Sub Products" is available, and clicking it adds the selected
sub products as non-editable lines. If the main product is removed,
its sub products are also removed automatically. On the sale order page,
a boolean field print_in_report determines whether sub products appear
in the printed sale order report and on the customer preview page,
if print_in report True then sub products show otherwise not.