[IMP] pricelist_price: add book price for account and sale
Allows users to view and compare the Unit price (actual selling price) and
Book price directly on the sale order lines.
Introduce a computed field book_price in sale.order.line and account.move.line.
book_price reflects the product price as per the selected pricelist, considering
product, quantity, and unit of measure.
Change sale_management in the manifest instead of the sale so we also access
the UI portion of module.

task- 4605742