[ADD] product_warranty: Added warranty functionality for eligible products
Implemented a feature to add a warranty for products marked as eligible.
Added a button to open a wizard for warranty selection.
Users can select the warranty from the wizard before applying it.
A sale order line is added after applying the warranty.
If the sale order line of the product is deleted, its corresponding warranty line is also removed.