[ADD] auction_real_estate:  Automated Real Estate Auction
Implemented the auction_real_estate module, first by adding an auction start
button and then introducing a radio button to enable auction-related fields.
When this button is enabled, all auction-related fields become visible.
Additionally, Implemented a restriction where offers cannot be manually
accepted if the auction is in the "Started" stage. Also developed an email
notification system so that users receive an email when their offer is either
accepted or rejected. Once the auction ends, the system updates the auction
details accordingly. To automate auction handling, Implemented a cron job
using the ir.cron model, which runs every five minutes to check whether the
auction time has ended. If the auction has ended, the system automatically
accepts the highest offer. In the Estate Settings, added two fields: Invoice
and Automated Auction. Enabling the Automated Auction option installs
the auction_real_estate module for the estate. On the properties page,
Implemented a filter to display only auction-related properties. On the
individual property page, added a "Create Offer" button and an auction
timer, which are displayed only when the auction is started. Created a
template for the "Create Offer" page, allowing users to place bids directly
from the website.