[ADD] deposit_rental: created new module 'deposit_rental'
Added a field in Rental configuration to set the deposit product.
Added a 'deposit_require' checkbox in the rental product form, along with a
'deposit_amount' field.
The 'deposit_amount' field is displayed only when the checkbox is selected.
The deposit amount is automatically updated when the deposit product is changed
in the settings.
When a rental product requiring a deposit is added to a Sale Order,
a deposit product is automatically added.
Additionally, the deposit product's quantity is updated whenever the linked
product's quantity in the Sale Order is modified.
Display deposit amount field in website, so that when the quantity is updated of
product the deposit amount should also be updated.
Also display the final amount in the website shop cart.