[ADD] rental_deposit: deposit in rental app
Added option to add deposit required and deposit amount field in rental product.
When in a sale order a rental product with this field ticked is added , a new
sale order line with deposit product which must have been configured previously
in the settings will also be added with unit price based on the quantity of the
rental product. In the ecommerce website too the deposit amount will be visible
and will change according to the quantity.

task-4632210