[ADD] product_category_global_info: add required attribute in global info page
Added two new options in the product category: one to show it in global info and
another for required attributes.
Updated the product category form to include these new options.
Ensured that when a product is added to a sale order, it checks if the category
should be shown in global info is true and retrieves the related details.
Created a new page in the sale order to display these details clearly.