[ADD] estate: base module handling properties management and salesflow
The main model of the module is estate.property, that contains all the basic data characterizing a property.
property.total_are is computed based on living_area and garden_area.
property keeps a link to res.user salesperson_id and to res.partner buyer_id.
roperty.state is originally set to 'new' and can be set to 'sold' or 'canceled' via two buttons. It is also updated based on property.offers_id.
Once a first property.offer is created, property.state is set to 'offer_received'.
Each time a new offer is created, it updates property.best_offer acordingly. A new offer's price cannot be lower than property.best_offer.
An offer can be validated or refused via two buttons. Once an offer is validated, all the other ones are set to 'refused', property.state is set to 'offer_accepted', property.selling_price is set to offer.price and property.buyer_id is set to offer.partner_id.
An offer lower than 90% of property.expected_price cannot be validated.
property.expected_price must have a strictly positive value.
property also have a list of tags and a type.
property.type keeps a list of all the properties and all the offers related to this specific type.