[IMP] estate: add extension classes
I've added extension classes in this commit, to extend the functionality of the base user model and be able to list all of the properties
he has listed in our module in the profile form view of the user. In addition to this, I've established the baseline of the estate_account model
to handle issuing an invoice for the buyers once a property has been sold.