[IMP] Added UI to the module.
Gave "write" and "create" access to all users of the database, to allow users to edit & create new records.

Created a new "ir.actions.act_window" action, that's associated with two views: list and form.

Created a menu item that triggers the action.

This is an exercise on the tutorial "Server Framework 101: Chapter 4"