[ADD] add estate app with empty view
[ADD] added views of adding, listing properties and search filters

[ADD] added 'offers' and 'other info' tabs in the property form and some new attributes

added the tags and property types, buyer, amd seller models and linked them in the property model.
also make view of offers list

[ADD] compute fields in property models

add the `total_area` and the `best_price` as computed attributes in the property modules and finish thier configuration

added `validity` and `date_deadline` in the offer model as inverse computed attributes and configure them

[FIX] styling issues

replaced harcooded font style with bootstrap classes, and make the code follow sake_case naming

[ADD] constrains to property fields and added some buttons for interactions

ad constrains in the `selling_price, expected_price` attributes in the property model to be positive and the `name, property_type_id` to be unique together

added consrains in the `price` attribute in the offer model to also be positive

removed unused imports

[ADD] interactive UI effects to the app

add colors to records according to its state, and hide some buttons if according to the step the property is in. added default filter to show available properties

[ADD] action button between property typers and offer

[ADD] invoicing model and model inheritance

inherit and add some vield to the user model

make property invoicing model that create invoice after selling property

[ADD] kanban view chapter 14

[FIX] warnings and style issues

fixed issue in the property type action button in the form view, the manifist data order was incorrect

fixed the warning of batch create in th property offer,

fixed some code style issues in the code

[FIX] handeled comments in the PR

- modify the way of getting the best price
- removed hard coded styles
- remove unwanted dependencies
- make code more clear