[REL] Estate: Added an offer and other infos tab
Added an offer tab as a one to many relationship to the property. Also added a view to see the price and expected prince and partner and orperty of the offer. In the other info tab; added a many to one relationship: added the fields buyer and seller, buyer is a partner and seller is a user. also added tags as a many to many relationship.

task-None
Fixes #None
Closes #None
opw-None