[IMP] Added multiple computed fields to the property & offer models.
The added computed fields for the `property` model are:
- best_price: highest amount offered for the property.
- total_area: sum of `living_area` and `garden_area`

The added computed fields for the `offer` model:
- deadline: sum of `create_date` and `validity_days`

Also added an "inverse" to the deadline field to update the validity when the deadline is edited, according to the create_date.

Updated the respective views to display the newly added fields.

This is an exercise on the tutorial "Server Framework 101: Chapter 8"