[IMP] : Adding CRUD conditions + Add herited notebook page Real Estate inside res.users
I added protection to avoid removing properties in state that aren't New or Cancelled.
I added an automatic status change for the properties when an Offer is created.

I added a rule to avoid accepting offers that are lower than previous ones.

I added a properties list inside the User form.