[IMP] awesome_owl : Adding Counter + Adding Card + Adding TodoList
I added a Counter Component with a button that I made reusable as a Subcomponent.

I added a Card component that can display content and add a show/hide option.

I added a TodoList with TodoItem.
I can create new todoItem using an input that detect Enter key.
The todoItem can be marked as completed.
It can also be remove.

I implement some Props validation for each props for each component.

I added a sum system for the Counter.