[IMP] estate : fixing creation bugs with multi offer creation
The check for the lowest price was made on the best price. But the best price was set before the end of the offers creation causing an issue when saving multiple offers.
I change the function to extract the old_offer and use the max_price of the offer as the validation.

The app can now create multiple offers at once without issues. They're validated in price ascending order.