[IMP] awesome_dashboard : Adding Dashboard Items (NumberCard and PieChartCard) + Implementing dashboard configuration
I added a Layout for the dashboard.

I added some buttons for quick navigations to leads and customers.

I added some dashboard items of 2 types : NumberCard and PieChartCard.

I added a service call to fetch statistics data and a cache system.
The data are fetched each 10 seconds.

All the dashboard is lazy_loaded.

The dashboard is generic so you can easily adapt it. It is possible to select the items you want to see using the configuration Dialog.

The texts are translatables.