[IMP] Awesome Dashboard 😎
A generic dashboard showing some statisitics fetched periodically from '/awesome_dashboard/statistics' controller.

The added dashboard items:
- Number of new orders this month
- Total amount of new orders this month
- Average amount of t-shirt by order this month
- Number of cancelled orders this month
- Average time for an order to go from ‘new’ to ‘sent’ or ‘cancelled’
- A piechart indicating the number of t-shirts ordered per t-shit size.

Added the option for the user to hide/show specific items using the cog wheel.

Note that the dashboard items are generic /and easy to extend, and are added to the registry "dashboard_items" within the "awesome_dashboard" category.

This is an exercise on the 'Discover the web framework: Chapter 2' tutorial.