[IMP] estate: enhanced estate module and create estate_account bridge module
added functionalities like inline list view, status bar widget, conditionally
hiding the buttons, color decoration of rows in list view, and manual ordering
of rows, enhanced the filter_search functionalities with the available as a
default field in the search view, adding the stats button of the offer.
Moreover, custom logic was added to CRUD operations,
and the res.user model was inherited.
Lastly, created the estate_account bridge module.