[IMP] estate: improvement estate module  and created estateAccount bridge module
Implement the status bar to view the application's status and how to work
in order and sequence in the particular field. Implementation of color option
and its attribute, which is working under different conditions, and also makes
a stat button, invisible options, and see the CRUD method, and inherit the model.
and created the estate account module.