[ADD] estate:  added access rights, field attributes and custom Views
Applied the security access to base users group using CSV.
Started working on UI:
Create the estate_property_views.xml file in the Views folder
Created an action for the model estate.property.
Created the estate_menus.xml
Created the three levels of menus for the estate.property

Added the appropriate default attributes so that:
the default number of bedrooms is 2
the default availability date is in 3 months

Added active and state field.

Defined a custom list view for the estate.property
Defined a custom form view for the estate.property
Defined a custom search view for the estate.property