[IMP] real_estate: give basic UI to real estate module and add security
In this tutorial, understand the concept of security and access rights
and add csv file with access right for the user. use Action and understand
how it will interact with UI. Menus link to actions, which in turn link to
models. Also learn about field behavior which control behavior of UI and
understand reserved filed. add list and form view and understand the
concept of search and different types of views.