[ADD] estate: added property type, Tags, offers and States
Added Property type model, created view and added it to menu

Added filed to Buyer and salesperson field to property model

Added Property tag model, created view and added it to menu

Created many2many relation between property tags and property model

Added Property offer model, created view and added property form page.

Added the total_area field to estate.property. It is defined as the sum of
 the living_area and the garden_area.

Added the best_price field to estate.property. It is defined as the highest
 (i.e. maximum) of the offers’ price.

Added validity and date_deadline to property offer

Defined inverse function so that the user can set either the date or the
validity.

Created an onchange in the estate.property model in order to set values for the
garden area (10) and orientation (North) when garden is set to True. When unset,
clear the fields.

Added the buttons ‘Cancel’ and ‘Sold’ to the estate.property model. A cancelled
property cannot be set as sold, and a sold property cannot be cancelled.

Added the buttons ‘Accept’ and ‘Refuse’ to the estate.property.offer model.

When an offer is accepted, set the buyer and the selling price for the
 corresponding property.