[IMP] real_estate: add test cases in real estate module
In this tutorial, get knowledge about runbot and create a test cases which
create offer for sold property and Sell a property with no accepted offers
implement this two test cases. also add reset of Garden Area and Orientation
when you uncheck the Garden checkbox this test case also implement.
this is useful for test some type of cases which can raise error.