[ADD] pricelist_price: add book price to sale order and invoice lines
Added a new field to capture book price in both sale order lines and
account move (invoice) lines.The 'Book Price' is dynamically calculated
for each order line based on the product,quantity, and the pricelist configured
on the parent sale order. If a pricelist is set, its rules are used to compute
the price for the product and quantity. If no pricelist is specified,
the product's standard list price is used as a fallback.This enhancement
improves pricing transparency and auditing.

task-4958277