[ADD] send_invoice: add send invoice automatically by email module
In this task,
add module which send invoice by email automatically.
in settings fill field enter the number of days to send email after the
invoice date. the purpose of the task is that do not need to send invoice
by manually it will automatically send by email depend upon
how many days will be selected.