[ADD] estate: Implement compute fields, onchange logic, actions, and constraints and UI Improvements
This update introduces several enhancements to the estate module. Implemented computed fields such as total_area and best_price to automatically calculate property details. An onchange method is added on the garden field to set default values for garden_area and garden_orientation when the garden option is enabled. New actions are added to mark properties as Sold or Cancelled, and offers can now be accepted or refused, with only one offer allowed to be accepted per property. When an offer is accepted, the buyer and selling price are updated on the property. SQL constraints ensure that expected_price and selling_price are always positive, also Property type should be unique and a Python constraint ensures that the selling price cannot be lower than 90% of the expected price.

Additionally,  done UI improvements for property type form view. A property list is now displayed inline on the property type form view showing name, expected price, and status for clarity. The property status field uses the statusbar widget to display states (New, Offer Received, Offer Accepted, Sold). The Sold and Cancel buttons are now conditionally hidden when the property is no longer editable, making the interface cleaner and preventing invalid actions.